Lag simulation based on 'laggy' listener, not sender

+++ core/src/otechniques/Config.java
@@ -30,9 +30,8 @@ public class Config {
 
 	public static final float INPUT_SPOOFING_CHANGE_FREQ = 0.5f;
 
-	public static final long CLIENT_PING = 200; // latency in packets sending,
-												// in milliseconds
-	public static long SERVER_PING = 200;
+	public static final int CLIENT_PING = 200; 
+	public static int SERVER_PING = 200;
 	public static final float PLAYER_STATE_SENDING_FREQUENCY = 1 / 60f;
 
 	public static boolean SERVER_RECONCILIATION = true;
